quizz/requirements.txt
pandas==1.5.3
Flask==2.2.2
python-dotenv==0.20.0
SQLAlchemy==1.4.46
psycopg2==2.9.5